backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Database ───────────────────────────────────────────────
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
databases==0.9.0
psycopg2-binary==2.9.9

# ── Supabase ───────────────────────────────────────────────
supabase==2.5.0

# ── Data Validation ────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Background Jobs ────────────────────────────────────────
redis==5.0.4
rq==1.16.2

# ── LLM & Research ─────────────────────────────────────────
anthropic==0.26.0
tavily-python==0.3.3

# ── Utilities ──────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.0
python-jose[cryptography]==3.3.0   # JWT verification